requirements.txt
numpy>=1.25
pandas>=2.0
PyYAML>=6.0
streamlit>=1.28.0
plotly>=5.18.0